requirements.txt
amqp==2.5.2
aniso8601==7.0.0
asgiref==3.2.7
babel==2.8.0
beautifulsoup4==4.7.1
billiard==3.6.3.0
boto3==1.13.7
botocore==1.16.7
braintree==3.49.0
cachetools==4.1.0
cairocffi==1.1.0
cairosvg==2.4.2
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-email-url==1.0.0
django==3.0.6
django-appconf==1.0.4
django-cache-url==3.1.2
django-countries==6.1.2
django-filter==2.2.0
django-graphql-jwt==0.3.0
django-js-asset==1.2.2
django-measurement==3.2.3
django-mptt==0.11.0
django-phonenumber-field==4.0.0
django-prices==2.2.0
django-prices-openexchangerates==1.1.0
django-prices-vatlayer==1.1.0
django-render-block==0.6
django-storages==1.9.1
django-templated-email==2.3.0
django-versatileimagefield==2.0
docutils==0.15.2
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
faker==4.0.3
freezegun==0.3.15
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-i18n-address==2.3.5
google-measurement-protocol==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
graphene==2.1.8
graphene-django==2.10.0
graphene-federation==0.1.0
graphql-core==2.3.1
graphql-relay==2.0.1
html-to-draftjs==1.0.1
html5lib==1.0.1
idna==2.9
jaeger-client==4.3.0
jmespath==0.9.5
jsonfield==3.1.0
kombu==4.6.8
lxml==4.5.0
markdown==3.2.2
maxminddb==1.5.4
maxminddb-geolite2==2018.703
measurement==3.2.0
mpmath==1.1.0
oauthlib==3.1.0
opentracing==2.3.0
phonenumberslite==8.12.3
pillow==7.1.2
prices==1.0.0
promise==2.3
protobuf==3.11.3
psycopg2-binary==2.8.5
purl==1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyjwt==1.7.1
pyphen==0.9.5
python-dateutil==2.8.1
python-magic==0.4.18
python-magic-bin==0.4.14; sys_platform == "win32"
pytz==2020.1
razorpay==1.2.0
redis==3.5.1
requests==2.23.0
rsa==4.0
rx==1.6.1
s3transfer==0.3.3
sentry-sdk==0.14.3
singledispatch==3.4.0.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
stripe==2.48.0
sympy==1.5.1
text-unidecode==1.3
threadloop==1.0.2
thrift==0.13.0
tinycss2==1.0.2
tornado==6.0.4
tqdm==4.46.0
typing==3.7.4.1
urllib3==1.25.9
uwsgi==2.0.18; sys_platform != "win32"
vine==1.3.0
weasyprint==51
webencodings==0.5.1
